rm useless dice set component this component has the same function than GameBoard component
