+++ src/components/GameBoard.jsx
@@ -1,15 +1,35 @@
 import React from 'react';
 import PropTypes from 'prop-types';
 
-import DiceSet from 'components/DiceSet';
+import uuid from 'react-uuid';
+
+import Dice from 'components/Dice';
 
 import 'styles/GameBoard.css';
 
-const GameBoard = ({ diceList, setDiceList, rolling }) => (
-  <div className="GameBoard">
-    <DiceSet diceList={diceList} setDiceList={setDiceList} rolling={rolling} />
-  </div>
-);
+const GameBoard = ({ diceList, setDiceList, rolling }) => {
+  const switchLock = (idx) => {
+    if (diceList[idx].validated) { return; }
+    const newDiceList = [...diceList];
+    newDiceList[idx].locked = !newDiceList[idx].locked;
+    setDiceList(newDiceList);
+  };
+
+  return (
+    <div className="GameBoard">
+      {diceList.map((dice, idx) => (
+        <Dice
+          key={uuid()}
+          idx={idx}
+          switchLock={switchLock}
+          locked={dice.locked}
+          value={dice.value}
+          rolling={rolling && !dice.locked}
+        />
+      ))}
+    </div>
+  );
+};
 
 export default GameBoard;
