--- src/styles/DiceSet.css
@@ -1,3 +1,0 @@
-.DiceSet{
-  display: flex;
-}
